+++ backend/controllers/PortfolioController.php
@@ -113,7 +113,8 @@ class PortfolioController extends Controller
     {
         $model = $this->findModel($id);
 
-        unlink(yii::getAlias('@root').'/uploads/portfolio/'.$model->image);
+        if($model->image)
+            unlink(yii::getAlias('@root').'/uploads/portfolio/'.$model->image);
 
         $model->delete();
